requirements.txt
tensorflow >= 2.1.0
google-cloud-storage >= 1.16.1
Pillow >= 6.0.0
boto3 >= 1.9.91
